requirements.txt
stem
Flask
Pillow
pyqrcode
pycryptodome
tk
PySocks
pypng